Worked on buttons that hide and display reply box
Joshua B. Adeniyi and Samuel Oketola helped me write the JavaScript that
would add the textboxs and display them when the buttons are clicked. I
continues by write the part that would hide the buttons when clicked but
display them when their alternate buttons are clicked.

Seeking for help when help is needed is helpful.